+++ eth/core/chain_context.go
@@ -37,7 +37,7 @@ type chainContext struct {
 
 // `GetHeader` returns the header for the given hash and height. This is used by the `GetHashFn`.
 func (cc *chainContext) GetHeader(_ common.Hash, height uint64) *types.Header {
-	header, _ := cc.blockchain.hp.GetHeaderByNumber(int64(height))
+	header, _ := cc.blockchain.bp.GetHeaderByNumber(int64(height))
 	return header
 }
