+++ host/cosmos/x/evm/plugins/block/plugin.go
@@ -22,44 +22,133 @@ package block
 
 import (
 	"context"
+	"math/big"
 
 	storetypes "cosmossdk.io/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	"pkg.berachain.dev/polaris/cosmos/x/evm/plugins"
+	"pkg.berachain.dev/polaris/eth/common"
 	"pkg.berachain.dev/polaris/eth/core"
+	coretypes "pkg.berachain.dev/polaris/eth/core/types"
 )
 
-// `Plugin` is the interface that must be implemented by the plugin.
+// TODO: change this.
+const bf = uint64(1)
+
 type Plugin interface {
 	plugins.BaseCosmosPolaris
 	core.BlockPlugin
+
+	// `SetQueryContextFn` sets the function used for querying historical block headers.
+	SetQueryContextFn(fn func(height int64, prove bool) (sdk.Context, error))
 }
 
-// `plugin` keeps track of polaris blocks via headers.
 type plugin struct {
 	// `ctx` is the current block context, used for accessing current block info and kv stores.
 	ctx sdk.Context
-	// `hp` represents the header plugin, used for accessing historical block headers.
-	hp core.HeaderPlugin
 	// `storekey` is the store key for the header store.
 	storekey storetypes.StoreKey
-	//  `offchainStore` is the offchain store, used for accessing offchain data.
-	offchainStore storetypes.CacheKVStore
+	// `getQueryContext` allows for querying block headers.
+	getQueryContext func(height int64, prove bool) (sdk.Context, error)
 }
 
-// `NewPlugin` creates a new instance of the block plugin from the given context.
-func NewPlugin(
-	hp core.HeaderPlugin, offchainStore storetypes.CacheKVStore, storekey storetypes.StoreKey,
-) Plugin {
+func NewPlugin(storekey storetypes.StoreKey) Plugin {
 	return &plugin{
-		hp:            hp,
-		offchainStore: offchainStore,
-		storekey:      storekey,
+		storekey: storekey,
 	}
 }
 
 // `Prepare` implements core.BlockPlugin.
 func (p *plugin) Prepare(ctx context.Context) {
 	p.ctx = sdk.UnwrapSDKContext(ctx)
+}
+
+// `BaseFee` returns the base fee for the current block.
+// TODO: implement properly with DynamicFee Module of some kind.
+//
+// `BaseFee` implements core.BlockPlugin.
+func (p *plugin) BaseFee() uint64 {
+	return bf
+}
+
+// `NewHeaderWithBlockNumber` builds an ethereum style block header from the current
+// context.
+func (p *plugin) NewHeaderWithBlockNumber(number int64) *coretypes.Header {
+	cometHeader := p.ctx.BlockHeader()
+	if cometHeader.Height != number {
+		panic("block height mismatch")
+	}
+
+	// We retrieve the `TxHash` from the `DataHash` field of the `sdk.Context` opposed to deriving it
+	// from solely the ethereum transaction information.
+	txHash := coretypes.EmptyRootHash
+	if len(cometHeader.DataHash) == 0 {
+		txHash = common.BytesToHash(cometHeader.DataHash)
+	}
+
+	parentHash := common.Hash{}
+	if number > 1 {
+		if header, err := p.GetHeaderByNumber(number - 1); err == nil {
+			parentHash = header.Hash()
+		} else {
+			panic("parent header not found")
+		}
+	}
+
+	return &coretypes.Header{
+		// `ParentHash` is set to the hash of the previous block.
+		ParentHash: parentHash,
+		// `UncleHash` is set empty as CometBFT does not have uncles.
+		UncleHash: coretypes.EmptyUncleHash,
+		// TODO: Use staking keeper to get the operator address.
+		Coinbase: common.BytesToAddress(cometHeader.ProposerAddress),
+		// `Root` is set to the hash of the state after the transactions are applied.
+		Root: common.BytesToHash(cometHeader.AppHash),
+		// `TxHash` is set to the hash of the transactions in the block. We take the
+		// `DataHash` from the `sdk.Context` opposed to using DeriveSha on the PolarisBlock,
+		// in order to include non-evm transactions block hash.
+		TxHash: txHash,
+		// We simply map the cosmos "BlockHeight" to the ethereum "BlockNumber".
+		Number: big.NewInt(cometHeader.Height),
+		// `GasLimit` is set to the block gas limit.
+		GasLimit: blockGasLimitFromCosmosContext(p.ctx),
+		// `Time` is set to the block timestamp.
+		Time: uint64(cometHeader.Time.UTC().Unix()),
+		// `BaseFee` is set to the block base fee.
+		BaseFee: big.NewInt(int64(p.BaseFee())),
+		// `ReceiptHash` set to empty. It is filled during `Finalize` in the StateProcessor.
+		ReceiptHash: common.Hash{},
+		// `Bloom` is set to empty. It is filled during `Finalize` in the StateProcessor.
+		Bloom: coretypes.Bloom{},
+		// `GasUsed` is set to 0. It is filled during `Finalize` in the StateProcessor.
+		GasUsed: 0,
+		// `Difficulty` is set to 0 as it is only used in PoW consensus.
+		Difficulty: big.NewInt(0),
+		// `MixDigest` is set empty as it is only used in PoW consensus.
+		MixDigest: common.Hash{},
+		// `Nonce` is set empty as it is only used in PoW consensus.
+		Nonce: coretypes.BlockNonce{},
+		// `Extra` is unused in Polaris, but can be used to store additional information, e.g.
+		// the host block hash.
+		Extra: []byte(nil),
+	}
+}
+
+// `blockGasLimitFromCosmosContext` returns the maximum gas limit for the current block, as defined
+// by either the block gas meter or the consensus parameters if the gas meter is not set or is an
+// InfiniteGasMeter. If neither the gas meter nor the consensus parameters are available, it
+// returns 0. This shouldn't be an issue in practice but we include this function for completeness
+// defensive programming purposes.
+func blockGasLimitFromCosmosContext(ctx sdk.Context) uint64 {
+	blockGasMeter := ctx.BlockGasMeter()
+	if blockGasMeter == nil || blockGasMeter.Limit() == 0 {
+		cp := ctx.ConsensusParams()
+		if cp == nil || cp.Block == nil {
+			return 0
+		}
+		return uint64(cp.Block.MaxGas)
+	}
+
+	return blockGasMeter.Limit()
 }
