+++ host/cosmos/x/evm/plugins/historical/genesis.go
@@ -18,7 +18,7 @@
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
 // TITLE.
 
-package header
+package historical
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -27,8 +27,8 @@ import (
 )
 
 // `InitGenesis` performs genesis initialization for the evm module. no-op.
-func (p *plugin) InitGenesis(ctx sdk.Context, genesisState *types.GenesisState) {}
+func (p plugin) InitGenesis(ctx sdk.Context, genesisState *types.GenesisState) {}
 
 // `ExportGenesis` returns the exported genesis state as raw bytes for the evm
 // module. no-op.
-func (p *plugin) ExportGenesis(ctx sdk.Context, genesisState *types.GenesisState) {}
+func (p plugin) ExportGenesis(ctx sdk.Context, genesisState *types.GenesisState) {}
