+++ host/cosmos/x/evm/plugins/historical/interfaces.go
@@ -18,7 +18,7 @@
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
 // TITLE.
 
-package block
+package historical
 
 type StakingKeeper interface {
 	// TODO: coinbase stuff
