+++ host/cosmos/x/evm/plugins/historical/plugin_test.go
@@ -18,7 +18,7 @@
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
 // TITLE.
 
-package block
+package historical
 
 import (
 	dbm "github.com/cosmos/cosmos-db"
